Move and add logout to NavigationBar
