+++ src/ECER.Clients.RegistryPortal/ecer.clients.registryportal.client/src/App.vue
@@ -50,20 +50,6 @@ export default defineComponent({
 
     return { userStore, oidcStore };
   },
-  methods: {
-    async logout() {
-      if (this.userStore.isAuthenticated && this.userStore.authority) {
-        if (this.userStore.authority == "bceid") {
-          await this.oidcStore.logout(this.userStore.authority);
-        }
-        if (this.userStore.authority == "bcsc") {
-          // BCSC does not support a session logout callback endpoint so just remove session data from client
-          await this.oidcStore.removeUser(this.userStore.authority);
-          this.userStore.$reset();
-        }
-      }
-    },
-  },
 });
 </script>
